+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ## Fixed
- - Fix docs [#21](https://github.com/rokwire/polls-building-block/issues/21)
+- Set appID and orgID for notifications [#26](https://github.com/rokwire/polls-building-block/issues/26)
+- Fix docs [#21](https://github.com/rokwire/polls-building-block/issues/21)
 
 ## [1.0.21] - 2022-08-11
 ### Fixed